Install vagrant and virtualbox if needed

+++ scripts/linux/setup-debian.sh
@@ -1,16 +1,20 @@
 #!/bin/sh
 
-echo "deb [arch=amd64] http://download.virtualbox.org/virtualbox/debian $(lsb_release -sc) contrib" >> /etc/apt/sources.list.d/virtualbox.list
+if ! command -v virtualbox &> /dev/null; then
+    echo "deb [arch=amd64] http://download.virtualbox.org/virtualbox/debian $(lsb_release -sc) contrib" >> /etc/apt/sources.list.d/virtualbox.list
 
-wget -q https://www.virtualbox.org/download/oracle_vbox_2016.asc -O- | apt-key add -
+    wget -q https://www.virtualbox.org/download/oracle_vbox_2016.asc -O- | apt-key add -
 
-apt-get update
-apt-get install -y virtualbox-6.0
+    apt-get update
+    apt-get install -y virtualbox-6.0
+fi;
 
-VAGRANT_VERSION="2.2.6"
-TEMP_DEB="$(mktemp)" &&
-wget -O "$TEMP_DEB" "https://releases.hashicorp.com/vagrant/${VAGRANT_VERSION}/vagrant_${VAGRANT_VERSION}_x86_64.deb" &&
-dpkg -i "$TEMP_DEB"
-rm -f "$TEMP_DEB"
+if ! command -v vagrant &> /dev/null; then
+    VAGRANT_VERSION="2.2.6"
+    TEMP_DEB="$(mktemp)" &&
+    wget -O "$TEMP_DEB" "https://releases.hashicorp.com/vagrant/${VAGRANT_VERSION}/vagrant_${VAGRANT_VERSION}_x86_64.deb" &&
+    dpkg -i "$TEMP_DEB"
+    rm -f "$TEMP_DEB"
 
-vagrant plugin install vagrant-hostsupdater
+    vagrant plugin install vagrant-hostsupdater
+fi;
